requirements.txt
flask
pysnmp
sqlalchemy
schedule
calendar

mysql-connector-python
